Revstring

+++ src/main/java/com/java/workshop/brocode/Maths.java
@@ -8,13 +8,13 @@ public class Maths {
             System.out.println(Math.E);
 
             result = Math.pow(2, 5); //power 
-            result = Math.abs(-5); //absolute value
+            result = (double) Math.abs(-5); //absolute value
             result = Math.sqrt(9); //square root
-            result = Math.round(3.14); //round of 
+            result = (double) Math.round(3.14); //round of 
             result = Math.ceil(3.14); //celling
             result = Math.floor(3.66); //floor
-            result = Math.max(10, 020); // max num
-            result = Math.min(10,20); //min num
+            result = (double) Math.max(10, 020); // max num
+            result = (double) Math.min(10,20); //min num
 
             System.out.println(result);
         }
